+++ Assets/CoreEngine/Test/Controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e659f9c8f533f4841b196599d1dd9cf1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
