+++ Assets/CoreEngine/Test/Controller/IsometricTransformer/TileTransformerTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d25a08aa7e7873144b70755d9ded01b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
