+++ Assets/CoreEngine/Test/Model.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5bb57df2891361143b3f20823d45d6ce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
