+++ Assets/CoreEngine/Test/Model/World.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 482b91a5ae86f5243a9bfe3b76c96d0a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
