+++ Assets/CoreEngine/Test/Model/World/WallTest.cs
@@ -1,0 +1,18 @@
+﻿using UnityEngine;
+using UnityEditor;
+using UnityEngine.TestTools;
+using NUnit.Framework;
+using System.Collections;
+
+using com.gStudios.isometric.model.world.wall;
+
+
+namespace Tests.Model.World {
+
+    public class WallTest {
+        IWall regularWall;
+        IWall nullWall;
+
+        // TODO: This tests need a level interface to substitute it.
+    }
+}
