--- Assets/CoreEngine/Test/WallTest.cs
@@ -1,15 +1,0 @@
-﻿using UnityEngine;
-using UnityEditor;
-using UnityEngine.TestTools;
-using NUnit.Framework;
-using System.Collections;
-
-using com.gStudios.isometric.model.world.wall;
-
-public class WallTest
-{
-    IWall regularWall;
-    IWall nullWall;
-
-    // TODO: This tests need a level interface to substitute it.
-}
